+++ python/GafferTest/TypedObjectPlugTest.py
@@ -36,6 +36,7 @@
 ##########################################################################
 
 import unittest
+import imath
 
 import IECore
 
@@ -215,6 +216,53 @@ class TypedObjectPlugTest( GafferTest.TestCase ) :
 
 		self.assertFalse( p1.acceptsChild( p2 ) )
 		self.assertRaises( RuntimeError, p1.addChild, p2 )
+
+	def testSerialisationWithoutRepr( self ) :
+
+		# Check that we can serialise plug values even when the
+		# stored `IECore::Object` does not have a suitable
+		# implementation of `repr()`.
+
+		v1 = IECore.TransformationMatrixfData(
+			IECore.TransformationMatrixf(
+				imath.V3f( 1, 2, 3 ),
+				imath.Eulerf(),
+				imath.V3f( 1, 1, 1 )
+			)
+		)
+
+		v2 = IECore.TransformationMatrixfData(
+			IECore.TransformationMatrixf(
+				imath.V3f( 4, 5, 6 ),
+				imath.Eulerf(),
+				imath.V3f( 2, 2, 2 )
+			)
+		)
+
+		with self.assertRaises( Exception ) :
+			eval( repr( v1 ) )
+
+		s = Gaffer.ScriptNode()
+		s["n"] = Gaffer.Node()
+		s["n"]["user"]["p1"] = Gaffer.ObjectPlug( defaultValue = v1, flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
+		s["n"]["user"]["p2"] = Gaffer.ObjectPlug( defaultValue = v2, flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
+
+		s2 = Gaffer.ScriptNode()
+		s2.execute( s.serialise() )
+		self.assertEqual( s2["n"]["user"]["p1"].defaultValue(), v1 )
+		self.assertEqual( s2["n"]["user"]["p2"].defaultValue(), v2 )
+		self.assertEqual( s2["n"]["user"]["p1"].getValue(), v1 )
+		self.assertEqual( s2["n"]["user"]["p2"].getValue(), v2 )
+
+		s["n"]["user"]["p1"].setValue( v2 )
+		s["n"]["user"]["p2"].setValue( v1 )
+
+		s2 = Gaffer.ScriptNode()
+		s2.execute( s.serialise() )
+		self.assertEqual( s2["n"]["user"]["p1"].defaultValue(), v1 )
+		self.assertEqual( s2["n"]["user"]["p2"].defaultValue(), v2 )
+		self.assertEqual( s2["n"]["user"]["p1"].getValue(), v2 )
+		self.assertEqual( s2["n"]["user"]["p2"].getValue(), v1 )
 
 if __name__ == "__main__":
 	unittest.main()